more subtle bug fixes
1. don't reset election timer on handling every vote response, only when vote granted
2. in handling append_entry_responses, if our term is stale we need to step down
and not mutate any other state on the node -- so add the early return
3. we have to be careful about updating match and nex idxs in the leader
based on the original message's args. using UDP sockets also complicates
this because msgs can arrive out of order. this is probably not 100% correct
and needs to be more carefully examined
4. appendEntriesResponse.term could be stale if we call _update_follower(),
which updates the term, AFTER the message object's already been created.
re-set the `term` on the meessage in case.
there's probably a better way to do this
5. when receive append entries as a candidate, we should step down